Preliminary working implementation of sparse mode for 1.1

+++ branches/sparse/src/chebyshev.c
@@ -1,8 +1,8 @@
 /* FILE : chebyshev.c
- * $Date:: 2011-01-31 10:08:19 +0600 #$
+ * $Date::                            $
  * Descr: routines for determining parameters of the chebyshev particles
  *
- * Copyright (C) 2011 ADDA contributors
+ * Copyright (C) 2011-2012 ADDA contributors
  * This file is part of ADDA.
  *
  * ADDA is free software: you can redistribute it and/or modify it under the terms of the GNU
@@ -16,10 +16,11 @@
  * You should have received a copy of the GNU General Public License along with ADDA. If not, see
  * <http://www.gnu.org/licenses/>.
  */
-
-#include <math.h>
-#include "const.h"
+#include "const.h" // keep this first
+// project headers
 #include "io.h"
+// system headers 
+#include <math.h>
 
 // LOCAL VARIABLES
 
@@ -56,7 +57,7 @@ static double Zfunc(double x,enum newt_func mode)
 		cn=cos(xn);
 		return (-e*n*sn*c-(1+e*cn)*s)/(2*e*n*sn*s-(1+e*(1+n*n)*cn)*c);
 	}
-	else LogError(ONE_POS,"Unknown mode %d for calling Zfunc",mode);
+	else LogError(ONE_POS,"Unknown mode %d for calling Zfunc",(int)mode);
 }
 
 //==========================================================
@@ -80,7 +81,7 @@ static double Xfunc(double x,enum newt_func mode)
 		cn=cos(xn);
 		return (e*n*cn*c-(1+e*sn)*s)/(-2*e*n*cn*s-(1+e*(1+n*n)*sn)*c);
 	}
-	else LogError(ONE_POS,"Unknown mode %d for calling Zfunc",mode);
+	else LogError(ONE_POS,"Unknown mode %d for calling Zfunc",(int)mode);
 }
 
 //==========================================================
